+++ src/MessageFlow/model/index.ts
@@ -1,5 +1,7 @@
 import * as MessageFlow from "./MessageFlow";
+import * as NodeType from "./NodeType";
 
 export {
-    MessageFlow
+    MessageFlow,
+    NodeType,
 }
